refactor(v0.6.1): split core/reasoning/modes/meta.py (31.6 KB → 4-file package)
Rule #5 (module size ≤ 20 KB) cleanup — the single grandfather entry
in tests/test_v06_module_size_gate.GRANDFATHERED was core/reasoning/
modes/meta.py at 31.6 KB. This PR splits the file into a meta/ package
of four sibling modules and removes the grandfather entry per the
anti-creep rule.

Split:
  meta/__init__.py — facade: re-exports handle_meta (1.1 KB)
  meta/_parse.py   — _THEME_PATTERNS + _classify_theme + _TYPE_LABELS
                     + _TYPE_FILTER_TOKENS + _THEME_FILTER_TOKENS +
                     _RECENT_TOKENS + _NARRATIVE_TOKENS +
                     _parse_meta_filter + _peek_entity_type (8.7 KB)
  meta/_degree.py  — _REL_TARGET_RE + _REL_TARGETID_RE +
                     _ENTITY_ID_RE + _norm_key + _peek_fm_section +
                     _build_degree_map (4.9 KB)
  meta/_render.py  — _render_type_detail + _render_theme_detail +
                     _render_recent_detail + _render_llm_narrative
                     (7.5 KB)
  meta/_handler.py — handle_meta orchestrator (12.8 KB)

All five files comfortably under the 20 KB ceiling. Public surface
preserved: `from core.reasoning.modes import handle_meta` (via
modes/__init__.py) keeps working byte-identically because the facade
re-exports the same symbol.

Logic byte-identical. No behaviour change — the split is mechanical
(symbol-level migration with imports threaded through). No changes
to public arguments, return shape, or dispatch order.

Verification:
  python -c "from core.reasoning.modes import handle_meta" → ok
  python -m unittest tests.test_v06_module_size_gate
    Ran 3 tests in 0.057s — OK
  python -m unittest tests.test_meta_mode
    13 tests; 1 pre-existing failure on main
    (FastPatternRoutingTests.test_specific_topic_does_NOT_route_to_meta)
    surfaced before this PR — IntentClassifier routes "OpenAI 최신
    뉴스" to meta via the "뉴스" theme token, independent of the
    split. Documented for a future routing-cleanup PR; NOT in this
    PR's scope (no change to routing logic).

Streak: core/retrieval + core/graph traversal 0 lines changed.
core/reasoning content is byte-identical (split only).

Quality delta: exempt (label: refactor) — mechanical symbol-level
split with no behavioural change. The measurement axis is
unaffected; multihop / RGB benchmarks don't route through
handle_meta.

Module-size gate (tests/test_v06_module_size_gate.py):
  - GRANDFATHERED dict is now empty — no core/**/*.py over 20 KB.
  - test_grandfather_entries_still_oversize anti-creep enforces
    that empty-list state.
